lecture11/assignment/requirements.txt
apache-airflow==2.9.1
requests==2.32.3
pandas==2.2.2
jsonschema==4.23.0